requirements.txt
llama-index
openai
pypdf
streamlit